requirements.txt
google_api_python_client==2.94.0
python-dotenv==1.0.0
textblob==0.17.1
